Add User, Tags, Comments CRUD

+++ app/Models/Post.php
@@ -58,7 +58,10 @@ class Post extends Model
     protected static function booted(): void
     {
         static::deleting(function (Post $post) {
-            Storage::disk('public')->delete($post->image);
+            if ($post->image && Storage::disk('public')->exists($post->image)) {
+                
+                Storage::disk('public')->delete($post->image);
+            }
         });
     }
 }
